% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chart_history.R
\name{add_chart_history}
\alias{add_chart_history}
\title{Adds chart history to a data frame}
\usage{
add_chart_history(df, artist, chart_url)
}
\arguments{
\item{df}{A data frame}

\item{artist}{Unquote name of the column that contains an artist name}

\item{chart_url}{Unquoted name of the column that contains the chart_url to search}
}
\description{
Add artist chart history to a data frame. It is recommended to use this in conjunction with \code{list_artist_charts()}
}
