% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_chart.R
\name{bb_chart}
\alias{bb_chart}
\title{Search Billboard Charts}
\usage{
bb_chart(chart_url, date = Sys.Date())
}
\arguments{
\item{chart_url}{The url indicating the chart to search. See \code{bbcharts::chart_table} for all available charts.}

\item{date}{Get chart data for a specificed date (YYYY-MM-DD).}
}
\description{
Retrieve information from a Billboard chart.
}
