% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decade_end.R
\name{decade_end_bb}
\alias{decade_end_bb}
\title{Search Decade End Billboard Charts}
\usage{
decade_end_bb(chart_url)
}
\arguments{
\item{chart_url}{The url indicating the chart to search. See \code{bbcharts::de_chart_table} for all available charts.}
}
\description{
Decade end Billboard chart
}
