% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_chart_history.R
\name{get_artist_chart_history}
\alias{get_artist_chart_history}
\title{Billboard Chart history for a given artist}
\usage{
get_artist_chart_history(artist, chart_url)
}
\arguments{
\item{artist}{The name of an artist}

\item{chart_url}{The url indicating the chart to search.}
}
\description{
List Billboard Chart history for a given artist and chart.
}
