% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_artist_charts.R
\name{list_artist_charts}
\alias{list_artist_charts}
\title{List Charts that an Artist Appears on}
\usage{
list_artist_charts(artist)
}
\arguments{
\item{artist}{The name of an artist}
}
\value{
A tibble with the name of the artist, the charts that they appear on, and the chart-url for said chart.
}
\description{
View all charts that an artist appears on.
}
