% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_end.R
\name{year_end_bb}
\alias{year_end_bb}
\title{Search Year End Billboard Charts}
\usage{
year_end_bb(chart_url, chart_year)
}
\arguments{
\item{chart_url}{The url indicating the chart to search. See \code{bbcharts::ye_chart_table} for all available charts.}

\item{chart_year}{Specify the year of interest.}
}
\description{
Retrieve a year end Billboardchart
}
